neon

+++ WorkSpaceManager/views.py
@@ -629,24 +629,32 @@ class WorkSpaceMemberManger(APIView):
 
 
     def post(self,request):
+
         request.data['workspace_id'] = request.user.current_workspace_id
-        permissions = request.data.get("permissions")
-        serializer_data = WorkSpaceMemberSerializer(data= request.data)
-        if serializer_data.is_valid():
-            new_member = serializer_data.save()
-            create_permission_for_member(member_id=new_member.id, permissions=permissions)
+        workspace_obj = WorkSpace.objects.get(id=request.user.current_workspace_id)
+        if workspace_obj.owner == request.user:
+            permissions = request.data.get("permissions")
+            serializer_data = WorkSpaceMemberSerializer(data= request.data)
+            if serializer_data.is_valid():
+                new_member = serializer_data.save()
 
-            return Response(status=status.HTTP_201_CREATED,data={
-                "status":True,
-                "message":"کاربر با موفقیت اضافه شد",
-                "data":serializer_data.data
+                create_permission_for_member(member_id=new_member.id, permissions=permissions)
+
+                return Response(status=status.HTTP_201_CREATED,data={
+                    "status":True,
+                    "message":"کاربر با موفقیت اضافه شد",
+                    "data":serializer_data.data
+                })
+            return Response(status=status.HTTP_400_BAD_REQUEST,data={
+                "status":False,
+                "message":"validation error",
+                "data":serializer_data.errors
             })
-        return Response(status=status.HTTP_400_BAD_REQUEST,data={
+        return Response(status=status.HTTP_403_FORBIDDEN,data={
             "status":False,
-            "message":"validation error",
-            "data":serializer_data.errors
+            "message":"you dont have permission dont try",
+            "data":{},
         })
-
 
     def put(self,request,member_id):
         member_obj = get_object_or_404(WorkspaceMember,id=member_id)
@@ -654,50 +662,63 @@ class WorkSpaceMemberManger(APIView):
         permission_list= data.get("permission_list")
 
         avatar_id =data.get("avatar_id",None)
+        workspace_obj = WorkSpace.objects.get(id=request.user.current_workspace_id)
+        if workspace_obj.owner == request.user:
+            for permission_item in permission_list:
+                permission_member_obj= MemberPermission.objects.get(id=permission_item['id'])
+                with open('main_perm.json', 'r', errors='ignore', encoding='UTF-8') as file:
+                    permission_type = permission_item["permission_type"]
 
-        for permission_item in permission_list:
-            permission_member_obj= MemberPermission.objects.get(id=permission_item['id'])
-            with open('main_perm.json', 'r', errors='ignore', encoding='UTF-8') as file:
-                permission_type = permission_item["permission_type"]
+                    data = json.load(file)
+                    for permission in data:
+                        if permission['permission_name'] == permission_member_obj.permission_name:
+                            for item in permission['items']:
 
-                data = json.load(file)
-                for permission in data:
-                    if permission['permission_name'] == permission_member_obj.permission_name:
-                        for item in permission['items']:
+                                # try:
+                                    view_permission_obj = permission_member_obj.view_names.get(view_name=item['view_name'])
 
-                            # try:
-                                view_permission_obj = permission_member_obj.view_names.get(view_name=item['view_name'])
+                                    for perm in item['permissions']:
 
-                                for perm in item['permissions']:
+                                        if perm['type'] == permission_type:
+                                            for method in perm['methods']:
+                                                method_obj = view_permission_obj.methods.get(method_name=method['method'])
 
-                                    if perm['type'] == permission_type:
-                                        for method in perm['methods']:
-                                            method_obj = view_permission_obj.methods.get(method_name=method['method'])
-
-                                            method_obj.is_permission = method['status']
-                                            method_obj.save()
+                                                method_obj.is_permission = method['status']
+                                                method_obj.save()
 
 
 
-                            # except:
-                            #     continue
+                                # except:
+                                #     continue
 
-                permission_member_obj.permission_type = permission_type
-                permission_member_obj.save()
-        if avatar_id:
-            main_file = MainFile.objects.get(id=avatar_id)
-            main_file.its_blong=True
-            main_file.save()
-            member_obj.user_account.avatar=main_file
-            member_obj.user_account.save()
-        return Response(status=status.HTTP_202_ACCEPTED,data={
-            "status":True,
-            "message":"با موفقیت بروزرسانی شد",
-            "data":WorkSpaceMemberSerializer(member_obj).data
+                    permission_member_obj.permission_type = permission_type
+                    permission_member_obj.save()
+            if avatar_id:
+                main_file = MainFile.objects.get(id=avatar_id)
+                main_file.its_blong=True
+                main_file.save()
+                member_obj.user_account.avatar=main_file
+                member_obj.user_account.save()
+            return Response(status=status.HTTP_202_ACCEPTED,data={
+                "status":True,
+                "message":"با موفقیت بروزرسانی شد",
+                "data":WorkSpaceMemberSerializer(member_obj).data
+            })
+        return Response(status=status.HTTP_403_FORBIDDEN,data={
+            "status":False,
+            "message":"you dont have permission dont try!!",
+            "data":{}
         })
 
 
     def delete(self,request,member_id):
-        workspace_member = get_object_or_404(WorkspaceMember,id=member_id)
-        workspace_member.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        workspace_obj  = WorkSpace.objects.get(id=request.user.current_workspace_id)
+        if request.user == workspace_obj.owner :
+            workspace_member = get_object_or_404(WorkspaceMember,id=member_id)
+            workspace_member.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(status=status.HTTP_403_FORBIDDEN, data={
+            "status": False,
+            "message": "you dont have permission dont try!!",
+            "data": {}
+        })
